new: Split filtering and completion interface
The filtering interface is between the Model API server and the
filter. The interface may be called to edit the user input or the
output.

The completion interface is between the Model API server and the LLMs. The LLMs should complete the requested text from the Model API server.

Note: The role of generatring prompt from user input is undefined
currentlly. It can be in a filter, in the pre-processing stage of
LLMs, or even the user can be in charge to prompt.