[Worker] fix: Handle interrupt during registeration
Original version won't handle any interrupt during the registeration
process.
This is due to we only have single thread and the registeration
subroutine will block it until the retry counter become zero.
The solution is making the registration process asynchronous and
scheduling this coroutine on start-up.
The coroutine will yield the CPU during I/O or backoffing. This solves
the stalling problem and the main thread is able to handle the interrupt
concurrently.