[Docker] fix: Dependency of scripts
This patch revised the structure of both docker image and docker
compose. The main contribution is to resolve the dependencies. The
scripts in the "Docker/" provides the solid context about making the
image of web server.
However, since the principle is lefting the "web" directory untouched,
the scripts are not merged into the "web" directory. This should be
fixed in the near future.