[DocQA] fix: Multiprocessing terminates server
The original server was randomly terminated after serving user requests.
After investigation, we found a similar issue in [1]. The root cause is
that when we use multiprocessing with uvicorn, the parent process and
the child process share the same signal file descriptor [2]. As a
result, when we send a terminate signal to the child process, the parent
process terminates too.

In this patch, we fixed this issue by properly multithreading with
ThreadPoolExecutor under the asyncio framework [3].

References:
[1]: https://github.com/tiangolo/fastapi/issues/3283
[2]: https://github.com/tiangolo/fastapi/issues/1487#issuecomment-1157066306
[3]: https://docs.python.org/3/library/asyncio-dev.html#concurrency-and-multithreading