[DocQA] fix: HTML extract precision
The experiment shows that, for news website, there's some irrelative text are retrieved.
This patch fixs this by forcing Trafilatura focusing on precision.
As a result, some short term will be filtered out and the long main text
should not be affected.
However, it's important to not that for some website, the recall may
decrease.