new: vector guard evaluation
This patch adds a set of simple scripts to evaluate the performance of
the vector guard detector.

Note that the architecture of the vector guard under evaluation differs
from the production one. Nevertheless, the working principles are the
same. These vector guards are based on semantics search and make
decisions with the similarity score.