[agent] fix: Duplicated warning message
This patch fixes the duplicated warning message issue.
The root cause is that the "func" variable in the decorator is static.
Thus, when we call the wrapped function multiple times, the internal
decorator will continue chaining the same function. As a result, the
safety guard can also be called multiple times.

The solution here is creating a local reference in the "wrap" function,
therefore the chaining will apply to different function references
instead of the same reference.