[detector] fix: Health checking
This patch fixes the docker health checking issue of the detector. The
root cause is that grpc_health_probe depends on the
grpc_health.v1.Health service. However, the detector does not provide
such a service. This patch fixes this by providing a health service that
always returns the detection service is serving.
Nevertheless, further health check of the detection service is needed
and should be done in the future.