[client] fix: performance
This patch fixes the performance of the post-filter.
The original post-filter checks the buffer at every iteration of the
underlying generator.
However, since the underlying buffer may receive multiple tokens during
the checking period, the unnecessary check is requested.
This patch fixes this by consuming the underlying generator as many as
possible in the given time budget.