[dispatcher] fix: Abortion of the executor
Although the executor accepts concurrent jobs by design, the current
implementation of the dispatcher doesn't allow this. Thus, this patch
aligns the scheduling process of the executor to the ordinary model
worker. As a result, the executor is unable to serve concurrent requests
currently. This needs refactoring in the future.