[snap] fix: Storage of multi-chat
Since the snap package is read-only, the "storage" directory in the
multi-chat can't work properly. This patch fixes this by copying the
"storage" directory to a writable area, or more specifically
${SNAP_DATA}. Moreover, to preserve the space, other referenced files
are symbolically linked. Another solution is binding the data directory
into the snap directory by mounting. But this solution needs the
additional privilege of mounting.