[multi-chat] fix: can't create room
The original "main" function in the RoomController was responsible for
three distinct purposes: (1) serving as the home page for the room tab,
(2) being the launch point for initiating new chatrooms, and (3)
displaying the chatroom itself. The refactor in commit acf875c, however,
separated the functionality of the home page for the room tab and
displaying the completed chatroom from the initiation of new chatrooms,
thereby preventing the creation of new chat rooms through this route.

This commit aligns roles (3) and (1), (2), into distinct routes;
distinguishingly, the paths for (1) and (2) are delineated on the basis
of session data.